+++ src/main/java/dev/willyelton/crystal_tools/common/config/VanillaSkillTreeLevel.java
@@ -1,0 +1,8 @@
+package dev.willyelton.crystal_tools.common.config;
+
+public enum VanillaSkillTreeLevel {
+    NETHERITE,
+    DIAMOND_NETHERITE,
+    ALL,
+    NONE
+}